Unique session IDs (and feature flags)
L2TPv3 session IDs have to be unique on the entire system, not just per tunnel.
The only reason that tunneldigger got away with using 1 for all sessions is that
older Linux kernels failed to properly check for duplicate session IDs.  That
got fixed by kernel commit dbdbc73b44782e22b3b4b6e8b51e7a3d245f3086.

This patch adds unique session IDs to tunneldigger in a backwards-compatible
way.  If both ends of the tunnel agree to use a unique session ID, they both
will use the tunnel ID as the session ID.  To manage this mutual agreement, we
introduce space for up to 32 feature flags.  Three messages are affected:

CONTROL_TYPE_USAGE optionally takes 4 additional bytes after the padding, used
to send the client's feature flags.  Brokers may skew the usage they report back
depending on client features.  Old clients will not send these bytes, which is
interpreted as all flags being 0.

CONTROL_TYPE_PREPARE is extended the same way, also sending the client's feature
flags, so that the server does not have to remember them.  These are the
features that the client is offering to be using for this session.  Old clients
will not send any flags, which tells the server that no features are supported.
Old servers will ignore flags if they received them.

CONTROL_TYPE_TUNNEL is extended the same way, reporting the features that are
actually going to be used for this session.  Typically, this will be the
client's features masked by what the server supports.  If a feature-aware client
talks to an old server, the flags are going to be missing, telling the client
that no features are going to be used in this session.

The first feature flag is 'unique session IDs'.  If enabled, the session ID of
each side of the connection is going to be the same as the tunnel ID.
Furthermore, the broker gains an option to report full usage for clients not
supporting unique session IDs, making sure they connect to other servers (if any
are available).